[escher/expression_field] cursorOffset must have signed type since it can be equal to -1

+++ apps/calculation/app.h
@@ -29,7 +29,7 @@ public:
     CalculationStore * calculationStore() { return &m_calculationStore; }
     char * cacheBuffer() { return m_cacheBuffer; }
     size_t * cacheBufferInformationAddress() { return &m_cacheBufferInformation; }
-    size_t * cacheCursorOffset() { return &m_cacheCursorOffset; }
+    int * cacheCursorOffset() { return &m_cacheCursorOffset; }
     Poincare::LayoutCursor::Position * cacheCursorPosition() { return &m_cacheCursorPosition; }
 
   private:
@@ -40,7 +40,7 @@ public:
     char m_calculationBuffer[k_calculationBufferSize];
     char m_cacheBuffer[EditExpressionController::k_cacheBufferSize];
     size_t m_cacheBufferInformation;
-    size_t m_cacheCursorOffset;
+    int m_cacheCursorOffset;
     Poincare::LayoutCursor::Position m_cacheCursorPosition;
   };
